Zoom 관련 함수 분리
Zoom 관련 함수 분리
